+++ lists/queue.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include "queue.h"
-#include "dlist.h"
+#include "dlist/dlist.h"
 
 struct _Queue{
   DList primero;
